Petrify Somewhat Functional
Colliding with the light zone near the monster will trigger the UI text
to change to "Oh dear". In order to allow the player to physcially move
into that area, we'll have to switch to using triggers. As note, the
player collision is acting a bit odd, and will have to be refined.