VR Adjustments
Teleportation no longer drops to the floor. May have some problems with teleporting and clipping if someone is standing right at the edge of the acceptable space for the kit. Otherwise fine.

Jenny Script updated to work on X for attack and Y for movement.

Teleport pads now show up red on the ground. Sprite wasn't changed though